requirements.txt
catkin-pkg
lark
empy
wheel
pyyaml
numpy
pytest
transformers
ros2_ws/src/exodapt_robot_pt
ros2_ws/src/actions